K8s source controller updates (#52)
* Copy ContainerSource conditions to KES

This makes the status of the KES more obvious by hoisting every
condition from the ContainerSource.

* Update the ContainerSource spec if needed

If the Spec of the KES changes, the ContainerSource needs to be updated
too. Use DeepEqual to only update when necessary.

* Add additional objects required for sample

Sample also needs a channel and RBAC to read events. A future addition
should allow the KES to take a service account to pass through to the
receive adapter pod.