GitHub Source controller (#36)
* Generate GitHubEventSource

This is the raw output of:
kubebuilder create api --group sources --version v1alpha1 --kind GitHubEventSource
hack/update-codegen.sh && hack/update-manifests.sh

Subsequent commits will clean up some of this generated output

* Split controller into provider + reconciler

Following the example of other new sources, this splits the generated
controller into a provider and reconciler and moves the rbac
annotations into a doc.go.

The generated `ReconcileGitHubEventSource` type was also renamed to
just `reconciler`.

* Setup a logging context

Replace usage of context.TODO() with context.Background() and create a
logger from that context.

* Rename GitHubEventSource to GitHubSource

There's no reason to have Event in the name here.

* Stub basic GitHubSource definition

This just adds spec.sink, status.conditions, and status.sinkUri and
corresponding tests, all copied from ContainerSource with removal of
unneeded Conditions.

* Stub in a basic GitHubSource reconciler

This switches to the controller sdk for the GitHubSource reconciler
and copies the relevant tests and sink reconciliation code from
ContainerSource.

There's an opportunity to pull some of this common code around
populating and testing the sink status out somewhere.

* Update generated files

* wip create ksvc for githubsources

* Test for repository, accessToken, secretToken when reconciling

* Update the githubsource types test for the new Valid condition

* Port GitHub receive adapter, more reconciling

This is getting close to working, but not yet.

* We don't actually need a separate serving client

* And we don't need serving client generated files either

* The GitHubSource controller needs RBAC for getting secrets

* Update deps/manifests after rebase

* Use the shared sinks.GetSinkURI for resolving sinks to URIs

* Add GitHubSource to kustomization.yaml

* Add servingv1alpha1 to the controller-manager's scheme

I'm not sure this is the best way to do this, but controller-manager
needs the serving.knative.dev v1alpha1 scheme registered and this
seemed like the appropriate place to do that.

* Minor changes based on manual e2e testing of GitHubSource

There are still several rough edges, but the happy path of
GitHubSource works well enough to deploy the receive adapter and
register a webhook at GitHub pointing to it.

It needs more tests, especially around the receive adapter since none
of that code is even manually tested yet.

* Add unit test for github webhook creation

* Update serving dep to 0.2.0

* cloudevents lib moved out of eventing, sink is a URI

The cloudevents lib has since been moved out of eventing, thus
removing the eventing dependency from GitHubSource.

The GitHubSource receive adapter was also treating the Sink URI as a
domain, but it is not just a bare domain now.

Other minor changes based on review feedback.

* GitHub receive adapter already knows what the event type is

There's no need to check the `X-GitHub-Event` header to determine
we're processing a pull request event. The GitHub webhooks library in
use already does that for us - only pull request events will be sent
to the pull request handler we register.

* Update the GitHubSource sample yaml

* Use OpenAPI schema for GitHubSource validation

With the right kubebuilder annotations, the OpenAPI schema gets
generated with the necessary fields marked as required. The MarkValid
status isn't removed yet - it will transform a bit to indicate when
referenced secrets don't yet exist.

* Surface invalid GitHubSource secret refs in the status

* Rename GitHubSource Repository to OwnerAndRepository

This clarifies that the field expects to be an owner/org and a
repository. The plumbing, comments, and tests are wired up to allow
this to also be a bare organization for organization-level events, but
so far pull requests are the only event type we handle end-to-end.

* Remove leftover note from kubebuilder

* Somehow eventing snuck back in Gopkg.lock during rebase

* Sinkable is now Addressable

* Pull GitHubSource receive_adapter into its own package

This pulls the majority of the logic out of main and into a new
`pkg/githubsource/receive_adapter.go`, along with accompanying
tests.

* Move Serving scheme registration to `pkg/apis/serving/v1alpha`

* Delete GitHub webhook when deleting a GitHubSource

* Oops - wrong Kubernetes "sets" import

* GitHubSource needs get/list/watch for Secrets

This isn't ideal, and I think this can be reduced back down to just
Get by bypassing the informer cache. But, for now, list and watch are
needed on Secrets as well.

* Surface missing sink and secret error messages via conditions