Use CR unstructured client for dynamic objects (#137)
* Use CR unstructured client for dynamic objects

No longer need to create dynamic clients. CR will do this for
us internally based on the TypeMeta of the unstructured object.

Also updates tests to call the interface Addressable instead of
Sinkable.

* Remove dynamic client from GetSinkURI tests

Can use the standard CR client now that it supports unstructured
objects.

* Make injection optional for SDK reconcilers

CR provides interfaces and helpers to detect when injection is possible.

* Add the Client injection interface back

This is intended to make forgetting to inject a client a compile
error instead of a runtime error.

* Remove unnecessary InjectConfig methods