Another attempt to add an E2E test (#178)
* Add E2E test

This is a copy of the eventing E2E test which checks if
KubernetesEventSource can deliver an event to a sink.

* Update deps for E2E test

Dep added a bunch of non-go vendored files from eventing/cmd/* for some
reason.

* Use the default integration test helpers

* Remove 20 second teardown sleep

This was fixed in Istio 1.0.

* Correct go_test_e2e parameter

* Copy upload-test-images.sh from eventing

* Fix a couple docs typos

* don't wait for gcppubsub to go away

It isn't installed by default for e2e tests.

* Use pkgTest namespace name everywhere

* Remove errant whitespace

* Copied e2e tests from eventing also need k8sevents image

This copies over the k8sevents image, some other necessary utility
classes, and updates the location the e2e tests look for images to the
published eventing-sources image path.

* Undo accidental reverting of #155

* Missed a couple of needed test helper files

* Remove the last reference to pkg eventing/test

Update the test import in k8s_events_test.go to eventing-sources/test
instead of eventing/test. This fixes an error around duplicated e2e
test flags and ensures we aren't reliant on implementation details of
the eventing repo's tests.

* Somehow I missed that the merge required updating codegen

* No longer a default.yaml but instead individual resources

* Run update-deps.sh after merge

* Adapt to changes in pkg/cloudevents

* Make test/README.md reflect the new reality

* Satisfy markdown linter's preference for ordered lists

* Update e2e-tests.sh to latest test-infra conventions

* 2018 -> 2019 copyright in all new files

* Cleanup a confusing comment in test/crd.go

* Switch to cloudevents/sdk-go for e2e test

* Source is not actually a string in cloudevents sdk-go

* A bit more DOCKER_REPO_OVERRIDE -> KO_DOCKER_REPO

I missed then when catching up on the latest test-infra changes for
this PR.

* e2e images publish with basenames now

* Duplicate less code from knative/pkg/test

* Add missing copyright header to test/e2e/e2e.go