Update event-sources to not use fallback logger (#360)
* Updated Event-Sources to not use fallback loggers

Fixes #332

Currently, event-sources don't properly set, or receive their loggers
from context.  Change this to set the initial logger in both the
event-source, as well as the underlying receiver.

* Drop configmap evironment variable logkey

The evironment variable wasn't set/used anywhere, drop it for a set
string.

* Update logger variable context setting

We weren't properly setting the logger context before, use it