Record more detail about ContainerSource reconcile (#186)
* Record more detail about ContainerSource reconcile

A SetSinkURIFailed event is recorded when sink uri failure occurs. This
should catch most of the misconfiguration errors.

A DeploymentGetFailed event is recorded when the deployment can't be
retrieved from the apiserver.

A DeploymentReady event is recorded when the deployment has at least 1
ready replica. This is the "Success" event.

Add error strings to all warning events and conditions.

* Add test for IsDeployed

* Add helper to mark deployed and record an event

* Rework IsDeployed

* Document setSinkURIArg

* Fix test missed during merge