EventTypes from listers. (#1016)
I'd removed this in my previous PR, but I realized looking through couchdb that the reason the eventTypeInformer was unused (other than events) is because the github source was breaking our pattern of listing/fetching through the lister instead of the client (which it gets through reconciler.Base).

This reverts the informer event bits, and makes the object fetches go through the lister for consistency with out best practices.